Deflake KsckRemoteTest.TestClusterWithLocation
Ksck::CheckMasterConsensus did not reset a boolean flag indicating
whether a master consensus conflict had been found. This meant that, if
an instance of the Ksck object was being used to check repeatedly for
master consensus conflict, it would fail to recognize if the conflict
went away. Indeed, this was happening in
RemoteKsckTest.TestClusterWithLocation and made the test about 1% flaky
on debug (8/1000 failed). I ran 1000 runs on debug with this fix and saw
no failures.

Also, the CheckMasterHealth and CheckMasterUnusualFlags calls
outside ASSERT_EVENTUALLY are unnecessary. I removed them.

Change-Id: I3cfffad1c997a820834fadb55491691ea29ffaa2
Reviewed-on: http://gerrit.cloudera.org:8080/13198
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>