[TTL cache] add periodic scrubbing thread
This patch adds an optional periodic scrubbing thread for the TTLCache.
If enabled, the thread invalidates expired entries, effectively evicting
them from the cache.  With this patch, the TTLCache evicts expired
entries not only during the insertion of new ones if the cache is at its
capacity, but also upon detection of an entry's expiry by the scrubbing
thread.

Added corresponding test coverage as well.

Change-Id: Ib3bbbcf6211930a5501a03d6a47f3d911118946c
Reviewed-on: http://gerrit.cloudera.org:8080/13201
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>