Deflake TestTimeouts.testLowTimeouts
It turns out that if a ConnectToCluster "RPC" results in an RPC to each
master and each of those RPCs times out, the ConnectToCluster RPC ends
up with a ServiceUnavailable status, not a TimedOut status. This updates
the test to reflect this.

Before this patch, I saw 58/200 failures; after it I saw zero. The
58/200 seems big compared to the frequency of this in precommits, etc,
but anyway this fix ought to help regardless of how frequent the
problem is.

Change-Id: I7d1a429fe5749c31495e0d10f74a0cfb1e623310
Reviewed-on: http://gerrit.cloudera.org:8080/13202
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>