RetryRule: don't report NTP desynchronization test failures
This mirrors the logic in build-support/report-test.sh which skips reporting
of a failed test if there was an NTP clock synchronization issue.

I went back and forth on whether to do this in RetryRule or ResultReporter,
but settled on RetryRule as it seemed like the better place for "policy". I
also figured spawning a subprocess is faster (and easier) than doing a
primitive grep via Java APIs.

I tested this by running a local test result server and injecting a failure
into a test, once with the magic desynchronization string, and once without.
I saw the failure get reported in the former and skipped in the latter.

Change-Id: Ibba9768775f225425e98c3c36733aefcc7a42a11
Reviewed-on: http://gerrit.cloudera.org:8080/13206
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>