[authz] validator for --sentry_service_rpc_addresses flag
This patch introduces a group validator for the
--sentry_service_rpc_addresses runtime flag.  It makes it necessary
to set a non-empty value for the --hive_metastore_uris flag if the
--sentry_service_rpc_addresses flag is set.  In other words, this patch
makes it explicitly impossible to run Kudu master for a hypothetical
configuration of Kudu+Sentry authz without HMS catalog.

That reflects the logical dependency of the Kudu+Sentry fine-grain
authz scheme on the HMS catalog.  The dependency is there by design.

This patchs also contains a test for the introduced flag validator.
As of now, all existing tests for Kudu+Sentry authz are run with
configuration where the integration with HMS catalog is enabled as well.

Change-Id: Iec0470f68e34edf72a9e8baf608eda1b83272921
Reviewed-on: http://gerrit.cloudera.org:8080/13227
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>