[catalog_manager] update criterion to choose authz provider
Prior to this change, the hypothetical case when Kudu master is given
Sentry server's RPC end-point but no HMS catalog RPC end-point, the
fine-grained authorization was not be enabled since the catalog
manager would use DefaultAuthzProvider instead of SentryAuthzProvider
for authz decisions.

Code-wise, SentryAuthzProvider does not explicitly depend on the HMS
catalog.  From that perspective, the decision whether to use
SentryAuthzProvider or DefaultAuthzProvider for authz decisions should
be gated only by the presence of the Sentry server's RPC end-point
in the corresponding runtime flags.

From the design perspective, Kudu+Sentry fine-grain authz scheme
logically depends on the integration with HMS catalog (that's why the
case described above is pure hypothetical one).  The logical dependency
will be addressed in a few follow-up changelists.

Change-Id: Iee1760a8fe6ffc9d6822db2472da5ddef78aec8d
Reviewed-on: http://gerrit.cloudera.org:8080/13223
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>