KUDU-2812: Fix error reporting in KuduRestore
Before this patch, calls to
`session.getPendingErrors.getRowErrors.length`
would clear the errors buffer before we called it again
to report the errors.

This patch ensures we capture the result of
`session.getPendingErrors` before checking the length.

Note: This patch also updates the message format for
both kudu-backup and kudu-spark.

Change-Id: Iadf3941614a4879a9f35d1df9ee0cea274711c94
Reviewed-on: http://gerrit.cloudera.org:8080/13244
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Kudu Jenkins