KUDU-2677 Raise default value for --tablet_history_max_age_sec
This raises the default --tablet_history_max_age_sec to 1 week, to
support incremental backup and the diff scans it relies on.

Raising the default will cause clients to be able to scan further into
the past, possibly into a region where the data has been GC'd. This
would result in no rows returned, which might be surprising to some
users. This condition will be temporary, and will end when the new
default period has passed. We don't expect this to be a problem because
we don't think many users are doing scans like this, and if they are
they probably adjusted --tablet_history_max_age_sec and so won't be
affected by the change in default. If they want to raise their own
setting for the flag to accomodate incremental backups then they will
be able to take steps to mitigate the above issue.

Change-Id: I6398b57ec1abcd12c59a3588dd1a61900c0ccdeb
Reviewed-on: http://gerrit.cloudera.org:8080/13265
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins