[master_path_handlers] fix build on macOS
A stopgap solution for EasyJson in case when 'size_t' is not the same
type as 'uint64_t'.  As an alternative, it would be possible to handle
that in a more generic way in easy_json.cc, but std::enable_if is not
applicable there.  The option of putting explicit instantiation of
template member function after its specialization and avoiding
a warning (-Winstantiation-after-specialization) might be a way to go.

This is a follow-up to fd6155fc00de48d17a65aafd79fbbad584a4f837.

After the patch mentioned above, but prior to this fix, an attempt to
build on macOS would fail with a linker error like below:

Undefined symbols for architecture x86_64:
  "kudu::EasyJson& kudu::EasyJson::operator=<unsigned long>(unsigned long)", referenced from:
      kudu::master::MasterPathHandlers::HandleTablePage(kudu::WebCallbackRegistry::WebRequest const&, kudu::WebCallbackRegistry::WebResponse*) in master_path_handlers.cc.o
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)

Change-Id: Ica806031be9af6cc31d9654d3eaa7ac98f616420
Reviewed-on: http://gerrit.cloudera.org:8080/13281
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>