KUDU-2763: Eliminate confusing log message
Whenever a new tablet is created, the first leader will always send its
first message with "preceding opid" set to (1,1). This generates a "log
matching property violated" message that can confuse operators, when the
behavior seen is actually what we expect. I wrapped the log output in a
conditional statement so that we don't output the "log matching property
violated" INFO message when the preceding_opid from the leader is equal to
(1,1) which indicates a new tablet replica, thus the message doesn't need
to be shown for this particular scenario.

Change-Id: I95dd73cb2876dc3def218d84316ca015ddc9f166
Reviewed-on: http://gerrit.cloudera.org:8080/13325
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>