hms: Adjust external table handling
This patch changes the handling of external HMS tables.
Currently we synchronize external HMS entries and
create all synchronized HMS entries as external.

With this patch, we only synchronize managed HMS
entries and create all entries as managed by default.
This allows external tables to work the same regardless
of HMS synchronization being enabled.

This patch touches many locations in the code because
the use and validation of the new behavior is tightly
coupled.

The changes include:
- Prevent altering table type via the KuduMetastorePlugin
- Ignore non-managed Kudu tables in the
  KuduMetastorePlugin
- Ignore non-managed Kudu tables in the
  hms_notification_log_listener
- Use managed tables by default
- Always use the new storage format for new entries
- Add the table_name property to the new storage format
- Adjust upgrade and downgrade logic for external tables
- Adjust HmsCatalog::PopulateTable logic for external
  tables
- Adjust the tooling logic to correctly handle external
  tables.

Note: The updates to tooling logic and testing was
minimal and will be improved in follow on patches.

Change-Id: I6b7495486fa86710afb670a8e12f72684ba7af8f
Reviewed-on: http://gerrit.cloudera.org:8080/13347
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>