external_mini_cluster: raise start process timeout to 60s
After commit 7a6df6aa1 changed SystemNtp::Init to wait using ntp_adjtime(),
a few test failures showed up with the following:

  14:01:46.947 [DEBUG - main] (MiniKuduCluster.java:165) Response: error {
    code: TIMED_OUT
    message: "failed to start masters: Unable to start Master at index 0: Timed out after 30.000ss waiting for process (/tmp/dist-test-taskOjnAIG/build/tsan/bin/kudu-master) to write info file (/tmp/dist-test-taskOjnAIG/test-tmp/mini-kudu-cluster9136683209203726578/master-0/data/info.pb)"
  }

This was because the kernel's clock lost synchronization for more than 30s,
at which point the minicluster gave up on the process even though
SystemNtp::Init was prepared to keep waiting.

Here's a bandaid: extend the minicluster's process starting timeout to 60s,
so that it matches the default value of --ntp_initial_sync_wait_secs. We
could tightly couple the two values, but I'm not sure that's worth doing.

Change-Id: Ib4172e5b45a325eab9c8387bcf25c106d31b3328
Reviewed-on: http://gerrit.cloudera.org:8080/13348
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>