authz: mark --trusted_user_acl as non-experimental
This is a follow-up to b03411c and 8005afc, wherein other authorization
flags were marked non-experimental.

This patch marks --trusted_user_acl as "sensitive" to match other ACLs.
I opted to not mark it "stable" just yet (the other ACL flags are
stable), since it's still a relatively young flag.

Change-Id: I828688edc574a828add50c2ad8b397a2edb070c4
Reviewed-on: http://gerrit.cloudera.org:8080/13351
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>