[tool] update type of loadgen's --num_rows_per_thread flag
This patch updates the type of the --num_rows_per_thread command-line
flag, changing it from unsigned.  With that, the value of 0 is not a
special value for this flag anymore.  Instead, any negative value
for this flag now means 'as many rows per thread as possible'.

With this change, it's now possible to use the 'kudu perf loadgen'
CLI tool to create empty tables, which is useful in various test
scenarios.

In addition, this patch contains a small clean-up on using
the --keep_auto_table loadgen's flag in some test scenarios.

Change-Id: I2712ac7678c9cfd9359629f11df3a86dd727997d
Reviewed-on: http://gerrit.cloudera.org:8080/13373
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>