KUDU-2809 (2/6): skip unobservable rows when iterating an MRS
An unobservable row is one whose entire lifespan is contained within the
time range of a diff scan. Such rows should be hidden from the diff scan
results as they represent "changes" that the client wouldn't expect to see.

This patch adds such a heuristic to the MRS iterator. It's relatively
straight-forward because the MRS only stores REDOs and always remembers its
rows' insertion timestamps. Delta stores are more complicated; a patch for
handling unobservable rows for them will follow.

Change-Id: I801991260749d1f810540cb32ec84c3ea6a02160
Reviewed-on: http://gerrit.cloudera.org:8080/13534
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins