KUDU-2514 Part 1: Support extra config for table.
We have thousands of tables in the Kudu cluster. It's hard to set a
uniform configuration for all tables. So, we support extra configuration
properties that are customizable on a per-table basis.

Note: In this patch, we implemented the framework, and the Java API. The
configuration item only validate framework & Java API. It doesn't actually
affect the tablet.

Change-Id: I0514507dca95602a97e954d1db464b907e073aae
Reviewed-on: http://gerrit.cloudera.org:8080/12468
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>