[java] Fix TestHybridTime
TestHybridTime tests client propagated timestamps when timestamp
assignment needs to increment the logical clock. However, sometimes
there are jumps in the logical values of the logical clock, which made
this test flaky because it counted on seeing specific numbers of rows at
"magic number" logical values. This adjusts the test to track the
logical values back-propagated from the server and tests the expected
state of the table against those values instead of constants.

I also simplified the test.

I ran the original version 1000 times in dist-test and saw 81 failures.
I ran the new version 1200 times and saw 0 failures.

Change-Id: I98a5bccf2fb366004659f3e6169b5f57f7ced485
Reviewed-on: http://gerrit.cloudera.org:8080/13397
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins