[backup] Add a tool to clean up old backups
This adds a tool that cleans up old backups. Eligibility for cleanup is
a property of the backup path, not the actual backup. There are two
conditions:
1. The path is not the restore path.
2. Every backup on the path is older than the expiration age, which the
   tool allows users to express in days. The default is 30 days.
If a path satisfies both conditions, the tool will delete its backups,
from most recent to least recent. This way, if the tool terminates
mid-delete, the next run of the tool can pick up where the previous one
left off. However, a backup on the restore path will not be deleted,
even if it lies on a path that is expired.

The tool has a dry run mode, a verbose mode, and a table name filter. It
does not normally output anything on success.

There are a couple of situations not currently handled by this tool:
1. Failed backups with no metadata.
2. Backup paths with no full backup at the beginning.
This is because the backup graph abstraction doesn't presently handle
these cases in a way the tool can use. The tool can be further improved
after a follow up to enhance the backup graph.

Change-Id: I4a4c4dde10ec3c3c9b2233ee2f7b13c46e963e39
Reviewed-on: http://gerrit.cloudera.org:8080/13405
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>