hms: have tools ignore other clusters
Currently, the HMS tooling will check and fix the master addresses field
of HMS entries based on strict string comparison, assuming that the HMS
entry found for a Kudu table _should_ belong to the Kudu cluster being
operated on.

However, it is possible for there to exist metadata for multiple Kudu
clusters in a single Hive Metastore instance. As such, tooling to check
and fix a Kudu cluster's HMS integration shouldn't necessarily affect
another's.

This patch addresses this by imbuing the HMS tooling with the following
behavior:
- If any master address in a given HMS entry coincides with the expected
  Kudu masters, fix them (as was the behavior before); this is common if
  adding new masters, or replacing dead masters.
- If none of the master addresses in the HMS entry coincide with those
  expected, the HMS entry may belong to another Kudu cluster. By
  default, the tools will ignore such HMS entries. The user can specify
  --noignore-other-clusters to have the tools check/fix such master
  addresses.

Change-Id: Ib4e2ad5835fd7fedd1e963d234b153c1df5f8766
Reviewed-on: http://gerrit.cloudera.org:8080/13411
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>