[backup] KUDU-2787 Allow single table failures for backup
This path enhances backup to be able to survive the failures of
individual tables. It adds a new option `--failOnFirstError` to the
backup job, which defaults to false, and which when true causes the
backup job to fail as soon as any table has an error (the previous
behavior).

A follow up will add a similar option to the restore job.

Change-Id: I4f3731f8df04791727c44af9ade9989f2cc6ef95
Reviewed-on: http://gerrit.cloudera.org:8080/13415
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>