[java] Deflake TestAsyncKuduSession with temporary workaround
TestAsyncKuduSession.test is kind of a mess. In particular, it has one
section where it tries to insert 21 ops into a session with 20 ops of
buffer space. It expects this to trigger a PleaseThrottleException.
However, it is inherently racy because the session may manage to flush
one or more buffers of ops to the server so no throttling is needed.
It's always been an icky test but it wasn't flaky until
2bf80b5147f09c980c2f2268ee3467a5b3305a57 cleaned up the session code a
bit. Now it's about 30% flaky in RELEASE mode according to the flaky
test dashboard. Until I can spend some time to redo the test and
understand exactly what about the previous change triggered the
flakiness, I'm turning off the check for an exception.

I also moved a debug log statement so there's less noise when calling
flush in non-AUTO_FLUSH_BACKGROUND modes.

Change-Id: I751f1ed3a151eee611a8b88c1527e12271024bf6
Reviewed-on: http://gerrit.cloudera.org:8080/13453
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>