[java] Improve TestAsyncKuduSession: Eliminate dead setup and other small things
This is the first in a series of patches improving TestAsyncKuduSession.
The main thing this patch does is complete the transition of the test to
using the KuduTestHarness. Some comments and extra setup and cleanup
code were unnecessary now that the test uses the harness. I also did
a few other small things.

Change-Id: I2bc9f32c8e4b8745726e18189c55687246c8bab9
Reviewed-on: http://gerrit.cloudera.org:8080/13458
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>