[java] Deflake TestAuthTokenReacquire.testBasicMasterOperations
This test was flaky because DDL operations aren't exactly-once
(KUDU-1527). The following could happen with table creation or deletion:
1. Thread A issues a CreateTable request for Table A.
2. The master receives the CreateTable request and gets to work creating
   Table A but has not responded to Thread A yet.
3. Thread B calls `dropConnectionsAndExpireTokens`, dropping connections
   to the master.
4. Thread A retries the CreateTable RPC, thinking it has failed because
   the connection was dropped.
5. The master responds with an error because the table exists.

This works around the problem by catching and ignoring exceptions
indicating that the table already exists in the case of table creation
and catching and ignoring exceptions that the table is deleted or does
not exist in the case of table deletion.

Before this patch, I saw 16/100 runs fail in dist-test against TSAN
binaries. With this patch, I saw 0/100 failures.

Change-Id: Ifef0a254c6eb1cadeb54e9cef52f78c7ccc7b9c5
Reviewed-on: http://gerrit.cloudera.org:8080/13461
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>