[java] Deflake TestSecurity.testImportExportAuthenticationCredentials
This test was flaky for the same reason that
SecureKuduSinkTest.testEventsWithShortTickets was flaky (as addressed
in 3cbc025869494be2a1fa91582dfabc95d7bc88f9): the test was using short
ticket lifetimes, and sometimes in TSAN the cluster would simply take
longer than the ticket lifetime to start, causing the minicluster
check to fail with "not authorized" errors. However, unlike the sink
test, I don't see a reason why this test needs to use short lifetimes.
It manually invalidates credentials and does not wait for expiration.
So, I simply unset the option making it use short ticket lifetimes.

Change-Id: Ifb9fd101deb82acc849ba5dd6dfe8a34a76eea03
Reviewed-on: http://gerrit.cloudera.org:8080/13466
Tested-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>