[spark tools] KUDU-2831: Deflake DistributedDataGeneratorTest.testGenerateRandomData
Oops. It seems it's not possible to retrieve an accumulator by name. We
were creating an new accumulator for collisions after the old one was
cleaned up because the job finished and all references to it ended. This
meant we never actually counted collisions, so when they happened our
expectation of the number of rows written was off. This fixes the
problem by plumbing a reference to the rows written and collisions
accumulators back to the test code.

Before this patch, I saw 12/1000 failures. After, 0/1000.

Change-Id: Ie2218328e400d76dbeab50cdb0d0241dfe56d20c
Reviewed-on: http://gerrit.cloudera.org:8080/13477
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>