[backup] Deflake TestKuduBackup.testRandomBackupAndRestore
TestKuduBackup.testRandomBackupAndRestore was flaky because it the
DataGenerator sometimes generated duplicate rows, which caused count
mismatches between the number of rows upserted and the number of rows
scanned from the table (particularly when the random schema had a single
INT8 primary key). This fix works around the problem by eliminating the
validation of backups against the generated rows, reducing the set of
validations to just the validation of the final state of the table
against the restored table.

Change-Id: Ic8500f884b05d09a8679863c380dd89f1675519f
Reviewed-on: http://gerrit.cloudera.org:8080/13499
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>